Refactor configuration setup to separate bot authentication from operational settings
This commit introduces a clear separation between
bot authentication parameters (token, admins) and
operational settings (like number of days,
timezone, etc.) by creating a new
BotOperationConfig class. This improves the
clarity and organization of the configuration
structure, making it easier to manage and
understand the different aspects of bot
configuration.

Changes to be committed:
	modified:   .gitignore
	modified:   app/bot.py
	modified:   app/config_data/config.py
	modified:   app/middlewares/__init__.py
	modified:   app/middlewares/config_middleware.py
	modified:   app/routers/user/all_bookings.py
	modified:   app/routers/user/book.py
	modified:   app/routers/user/cancel_bookings.py
	modified:   app/routers/user/my_bookings.py
	modified:   poetry.lock
	modified:   pyproject.toml